+++ internal/provider/provider_next.go
@@ -132,6 +132,7 @@ func (p *frameworkProvider) DataSources(ctx context.Context) []func() datasource
 func (p *frameworkProvider) Resources(ctx context.Context) []func() resource.Resource {
 	return []func() resource.Resource{
 		NewAuditTrailTokenResource,
+		NewOrganizationDefaultSettings,
 		NewOrganizationRunTaskGlobalSettingsResource,
 		NewOrganizationRunTaskResource,
 		NewRegistryGPGKeyResource,
@@ -150,5 +151,7 @@ func (p *frameworkProvider) Resources(ctx context.Context) []func() resource.Res
 func (p *frameworkProvider) EphemeralResources(ctx context.Context) []func() ephemeral.EphemeralResource {
 	return []func() ephemeral.EphemeralResource{
 		NewAgentTokenEphemeralResource,
+		NewOrganizationTokenEphemeralResource,
+		NewTeamTokenEphemeralResource,
 	}
 }