+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 FEATURES:
 
+* **New Ephemeral Resource:** `tfe_organization_token` is a new ephemeral
+  resource for creating and managing organization tokens, by
+  @ctrombley [#1616](https://github.com/hashicorp/terraform-provider-tfe/pull/1616)
+
+* **New Ephemeral Resource:** `tfe_team_token` is a new ephemeral
+  resource for creating and managing team tokens in an organization, by
+  @shwetamurali and @ctrombley [#1628](https://github.com/hashicorp/terraform-provider-tfe/pull/1628)
+
 * `r/tfe_workspace_run`: Allow configuration without `apply` and `destroy` blocks, by @tmatilai [#1604](https://github.com/hashicorp/terraform-provider-tfe/pull/1604)
 
 ## v.0.64.0